- refactoring the cli builder a little
