+++ dyd/roots/source/dryad/dyd/assets/cli-builder/app.go
@@ -217,36 +217,3 @@ func (a *app) Run(appargs []string, w io.Writer) int {
 func (a *app) Usage(invocation []string, w io.Writer) error {
 	return Usage(a, invocation, w)
 }
-
-type arg struct {
-	key      string
-	descr    string
-	at       ArgType
-	optional bool
-}
-
-func (a arg) Key() string {
-	return a.key
-}
-
-func (a arg) Description() string {
-	return a.descr
-}
-
-func (a arg) Type() ArgType {
-	return a.at
-}
-
-func (a arg) Optional() bool {
-	return a.optional
-}
-
-func (a arg) WithType(at ArgType) Arg {
-	a.at = at
-	return a
-}
-
-func (a arg) AsOptional() Arg {
-	a.optional = true
-	return a
-}
